requirements.txt
fastapi
uvicorn[standard]
google-generativeai
supabase
pydantic